requirements.txt
streamlit 
bcrypt 
pillow